refactor(skills): extract external runtime sync
